+++ src/Resources/contao/languages/de/tl_linktracker_items.php
@@ -1,0 +1,19 @@
+<?php 
+
+/**
+ * Contao Open Source CMS
+ * Copyright (C) 2005-2014 Leo Feyer
+ *
+ */
+
+// Listenansicht
+$GLOBALS['TL_LANG']['tl_linktracker_items']['new'] = array('Neuer Aufruf', 'Neuen Aufruf eintragen');
+
+// Eingabemaske
+$GLOBALS['TL_LANG']['tl_linktracker_items']['tracker_legend'] = 'Aufrufdetails';
+$GLOBALS['TL_LANG']['tl_linktracker_items']['clickTime'] = array('Aufrufzeit', 'Zeitpunkt des Aufrufs');
+$GLOBALS['TL_LANG']['tl_linktracker_items']['ip'] = array('IP-Adresse', 'IP-Adresse des Besuchers');
+$GLOBALS['TL_LANG']['tl_linktracker_items']['browser'] = array('Browserkennung', 'Browserkennung des Besuchers');
+
+$GLOBALS['TL_LANG']['tl_linktracker_items']['publish_legend'] = 'Veröffentlichung';
+$GLOBALS['TL_LANG']['tl_linktracker_items']['published'] = array('Veröffentlicht', 'Aufruf veröffentlicht');