+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Linktracker
 
+## Version 1.0.1 (2021-11-30)
+
+* Add: Funktion is_bot zur Bot-Erkennung
+* Add: Statistikanzeige gesamt und letzte 4 Tage in Übersicht
+* Add: Übersetzungen tl_linktracker_items
+
 ## Version 1.0.0 (2021-11-30)
 
 * Ausbau der Version bis zur Einsatzreife